requirements.txt
numpy==2.2.2
Pillow==11.1.0